Type cast the returned value to float (task #2857)
This is useful for cases where the value is null or empty.

+++ src/FieldHandlers/DecimalFieldHandler.php
@@ -16,7 +16,7 @@ class DecimalFieldHandler extends BaseFieldHandler
      */
     public function renderValue($table, $field, $data, array $options = [])
     {
-        $result = filter_var($data, FILTER_SANITIZE_NUMBER_FLOAT, FILTER_FLAG_ALLOW_FRACTION);
+        $result = (float)filter_var($data, FILTER_SANITIZE_NUMBER_FLOAT, FILTER_FLAG_ALLOW_FRACTION);
 
         $args = $this->_getDbColumnArgs($table, $field);
